Design and add agent settings modal
