+++ frontend/src/app/(dashboard)/overview/page.test.tsx
@@ -2,7 +2,15 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { renderWithProviders, screen, waitFor, userEvent } from '@/test/test-utils';
 import Overview from './page';
 
-const { loginMock, sentryLoginMock, codexStatusMock, codexInitiateMock } = vi.hoisted(() => ({
+const {
+  loginMock,
+  sentryLoginMock,
+  codexStatusMock,
+  codexInitiateMock,
+  settingsGetMock,
+  settingsUpdateMock,
+  agentDefaultsMock,
+} = vi.hoisted(() => ({
   loginMock: vi.fn(),
   sentryLoginMock: vi.fn(),
   codexStatusMock: vi.fn().mockResolvedValue({ data: { status: 'pending' } }),
@@ -13,6 +21,17 @@ const { loginMock, sentryLoginMock, codexStatusMock, codexInitiateMock } = vi.ho
       expires_in: 900,
     },
   }),
+  settingsGetMock: vi.fn().mockResolvedValue({
+    data: {
+      name: 'Test Org',
+      settings: {
+        default_agent_type: 'codex',
+        agent_config: {},
+      },
+    },
+  }),
+  settingsUpdateMock: vi.fn().mockResolvedValue({ data: {} }),
+  agentDefaultsMock: vi.fn().mockResolvedValue({ data: {} }),
 }));
 
 vi.mock('@/lib/api', () => ({
@@ -25,6 +44,11 @@ vi.mock('@/lib/api', () => ({
       status: codexStatusMock,
       initiate: codexInitiateMock,
     },
+    settings: {
+      get: settingsGetMock,
+      update: settingsUpdateMock,
+      getAgentDefaults: agentDefaultsMock,
+    },
   },
 }));
 
@@ -34,6 +58,9 @@ describe('OverviewPage', () => {
     sentryLoginMock.mockReset();
     codexStatusMock.mockClear();
     codexStatusMock.mockResolvedValue({ data: { status: 'pending' } });
+    settingsGetMock.mockClear();
+    settingsUpdateMock.mockClear();
+    agentDefaultsMock.mockClear();
   });
 
   it('starts GitHub onboarding directly from the dashboard', async () => {
@@ -71,8 +98,36 @@ describe('OverviewPage', () => {
       expect(screen.getByText('Connect your coding agent')).toBeInTheDocument();
     });
 
-    expect(screen.getByText('Sign in with ChatGPT')).toBeInTheDocument();
-    expect(screen.getByRole('link', { name: 'Settings' })).toHaveAttribute('href', '/settings');
+    expect(screen.getByRole('button', { name: 'Sign in with ChatGPT' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'Settings' })).toBeInTheDocument();
+  });
+
+  it('opens agent settings modal from setup card and saves updates', async () => {
+    const user = userEvent.setup();
+    renderWithProviders(<Overview />);
+
+    await waitFor(() => {
+      expect(screen.getByRole('button', { name: 'Settings' })).toBeInTheDocument();
+    });
+
+    await user.click(screen.getByRole('button', { name: 'Settings' }));
+
+    await waitFor(() => {
+      expect(screen.getByText('Edit agent settings')).toBeInTheDocument();
+    });
+
+    expect(screen.getByText('For gpt-5.3-codex model access.')).toBeInTheDocument();
+    expect(screen.getByText('Recommended')).toBeInTheDocument();
+
+    await user.clear(screen.getByLabelText('Model'));
+    await user.type(screen.getByLabelText('Model'), 'codex-mini');
+    await user.click(screen.getByRole('button', { name: 'Save changes' }));
+
+    await waitFor(() => {
+      expect(settingsUpdateMock).toHaveBeenCalledTimes(1);
+    });
+
+    expect(screen.getByRole('link', { name: 'Open advanced agent settings' })).toHaveAttribute('href', '/settings/agents');
   });
 
   it('shows the AgentSetupCard as connected when auth status is completed', async () => {
